refactor(native): keep data-plane modules bounded
